// frontend/src/types/generated/HomeDashboardResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HomeDashboardPreference } from "./HomeDashboardPreference";
import type { HomeDashboardWidget } from "./HomeDashboardWidget";
import type { HomeDashboardWidgetKind } from "./HomeDashboardWidgetKind";

export type HomeDashboardResponse = { preferences: HomeDashboardPreference, available_widgets: Array<HomeDashboardWidgetKind>, widgets: Array<HomeDashboardWidget>, };